navigation items typescript
